cleanup crew

+++ src/general/store/ArriveIntervalCalculator.java
@@ -3,7 +3,7 @@ package general.store;
 public class ArriveIntervalCalculator {
     ExponentialRandomStream random;
 
-    public ArriveIntervalCalculator(double arriveInterval, int seed) {
+    public ArriveIntervalCalculator(double arriveInterval, long seed) {
         random = new ExponentialRandomStream(arriveInterval, seed);
     }
