+++ src/general/store/CashierSpeedCalculator.java
@@ -3,7 +3,7 @@ package general.store;
 public class CashierSpeedCalculator {
     UniformRandomStream random;
 
-    public CashierSpeedCalculator(double cashierMin, double cashierMax, int seed) {
+    public CashierSpeedCalculator(double cashierMin, double cashierMax, long seed) {
         random = new UniformRandomStream(cashierMin, cashierMax, seed);
     }
