+++ src/general/store/PickingTimeCalculator.java
@@ -3,7 +3,7 @@ package general.store;
 public class PickingTimeCalculator {
     UniformRandomStream random;
 
-    public PickingTimeCalculator(double pickingMin, double pickingMax, int seed) {
+    public PickingTimeCalculator(double pickingMin, double pickingMax, long seed) {
         random = new UniformRandomStream(pickingMin, pickingMax, seed);
     }
